Fix missing shell initialisation when doing a pipe straight away
```
- microstep:
  - local2in
    - /tmp/local
    - /tmp/in
```
was working because the alias actually translates to:
```
- microstep:
  - exec_in: mkdir -p /tmp
  - pipe:
      - exec_local: cat /tmp/local
      - exec_in: cat > /tmp/in
```
where `exec_in` actually take care of starting the shell of the `in`
context.

When doing the pipe without an `exec_in` beforehand, e.g.
```
- microstep:
  - pipe:
      - exec_local: cat /tmp/local
      - exec_in: cat > /tmp/in
```
it failed with a `fatal error` because the shell of the `in` context was
never started (=> stderr is nil).