Add option to enable creating checkpoint on all microsteps
By default (or if --enable-checkpoint is set to first-microstep),
kameleon only creates a checkpoint on the first not yet checkpointed
microstep of each macrostep:
- if no checkpoint exists for a macrostep, kameleon checkpoints its
first microstep only.
- if a checkpoint already exists for a microstep (and not later
macrostep is checkpointed), then the next microstep of the current
macrostep is run and checkpointed

If --enable-checkpoint is set to all-microsteps, then kameleon creates
checkpoints for all microsteps after the latest checkpointed macrostep.